changed directory name so it thinks everything is new...
