--- modules/debugmsgs.py
@@ -1,30 +1,0 @@
-# Colors yay
-import os, sys, platform, traceback
-
-scriptFile = None
-def init(file):
-    global scriptFile
-    scriptFile = file
-
-# Functions for different output messages (they are colorfull so they are easy to spot)
-def debugMsg(message):
-    # Print message with different formatting based on operating system
-    print('DEBUG: ' + str(message))
-
-def successMsg(message):
-    # Print message with different formatting based on operating system
-    print('SUCCESS: ' + str(message))
-
-def errorMsg(message, error, optionalScriptFile=None):
-    # Print message with different formatting based on operating system
-    if optionalScriptFile != None:
-        print('ERROR: ' 
-            + str(message) + '\n\n\t' 
-            + f'{optionalScriptFile}' + '\n\t> '+ 
-            str(error) + f' -> [Line: {error.__traceback__.tb_lineno}]')
-    else:
-        print('ERROR: ' 
-            + str(message) + '\n\n\t' 
-            + f'{scriptFile}' + '\n\t> '+ 
-            str(error) + f' -> [Line: {error.__traceback__.tb_lineno}]')
-    sys.exit(1)
